refactor: rename content-script -> content-scripts
