+++ src/entries/content-scripts/autoRefresh.ts
