+++ src/entries/content-scripts/descriptionColor.ts
