+++ src/entries/content-scripts/index.ts
