+++ src/entries/content-scripts/main.ts
