+++ src/entries/content-scripts/openInTabs.ts
