+++ src/entries/content-scripts/removeTableRowHighlight.ts
