+++ src/entries/content-scripts/search.ts
