+++ src/entries/content-scripts/updateFavicon.ts
